--- models/macro.py
@@ -1,13 +1,0 @@
-from ..app import db
-
-
-class Macro(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-    date = db.Column(db.Date, nullable=False)
-    protein = db.Column(db.Float, nullable=False)
-    carbohydrates = db.Column(db.Float, nullable=False)
-    fat = db.Column(db.Float, nullable=False)
-
-    def __repr__(self):
-        return '<Macro %r>' % self.id
